% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-check.R
\name{collate_editor_check}
\alias{collate_editor_check}
\title{Collate list of checks to single concatenated character string}
\usage{
collate_editor_check(checks)
}
\arguments{
\item{checks}{Result of \code{pkgcheck::pgkcheck} function}
}
\value{
Single character
}
\description{
Collate list of checks to single concatenated character string
}
\note{
Exported only for access by plumber; not intended for general external
usage.
}
\seealso{
Other main: 
\code{\link{editor_check}()},
\code{\link{serve_api}()}
}
\concept{main}
