% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch.R
\name{cran_to_list}
\alias{cran_to_list}
\title{Convert names of CRAN packages to markdown-formatted list.}
\usage{
cran_to_list(matches, n = 5L)
}
\arguments{
\item{matches}{A \pkg{pkgmatch} \code{data.frame} object with columns of
("package", "version", "rank").}

\item{n}{Number of matches to return in list.}
}
\description{
This function is exported because it needs to be called in the main plumber
endpoint function.
}
