% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_repo.R
\name{dl_gh_repo}
\alias{dl_gh_repo}
\title{Download a GitHub repo to local cache}
\usage{
dl_gh_repo(u, branch = NULL)
}
\arguments{
\item{u}{URL of GitHub repository}

\item{branch}{Checkout specified (non-default) branch of repo.}
}
\value{
Path to locally cached '.zip' version of repository
}
\description{
Download a GitHub repo to local cache
}
\seealso{
Other github: 
\code{\link{post_to_issue}()}
}
\concept{github}
