% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-check.R
\name{editor_check}
\alias{editor_check}
\title{Body of main 'editorcheck' response}
\usage{
editor_check(repourl, repo, issue_id, post_to_issue = TRUE)
}
\arguments{
\item{repourl}{The URL for the repo being checked, potentially including full
path to non-default branch.}

\item{repo}{The 'context.repo' parameter defining the repository from which
the command was invoked, passed in 'org/repo' format.}

\item{issue_id}{The id (number) of the issue from which the command was
invoked.}

\item{post_to_issue}{Integer value > 0 will post results back to issue (via
'gh' cli); otherwise just return character string with result.}
}
\value{
If \code{!post_to_issue}, a markdown-formatted response body from static
package checks, otherwise URL of the issue comment to which response body has
been posted.
}
\description{
Body of main 'editorcheck' response
}
\seealso{
Other main: 
\code{\link{collate_editor_check}()},
\code{\link{serve_api}()}
}
\concept{main}
