% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_branch_from_url}
\alias{get_branch_from_url}
\title{Get branch from a GitHub URL if non-default branch specified there}
\usage{
get_branch_from_url(repourl)
}
\arguments{
\item{repourl}{Potentially with "/tree/branch_name/sub-directory" appended}
}
\value{
Branch as single string.
}
\description{
Get branch from a GitHub URL if non-default branch specified there
}
\seealso{
Other github: 
\code{\link{dl_gh_repo}()},
\code{\link{get_subdir_from_url}()},
\code{\link{post_to_issue}()},
\code{\link{url_is_r_pkg}()}
}
\concept{github}
