% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_subdir_from_url}
\alias{get_subdir_from_url}
\title{Return sub-directory from URL if present. This function is also intended to
test whether packages are in sub-directories (issue #64)}
\usage{
get_subdir_from_url(repourl)
}
\arguments{
\item{repourl}{Potentially with "/tree/branch_name/sub-directory" appended}
}
\description{
Return sub-directory from URL if present. This function is also intended to
test whether packages are in sub-directories (issue #64)
}
\seealso{
Other github: 
\code{\link{dl_gh_repo}()},
\code{\link{get_branch_from_url}()},
\code{\link{post_to_issue}()},
\code{\link{url_is_r_pkg}()}
}
\concept{github}
