% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer-review-badge.R
\name{has_readme_badge}
\alias{has_readme_badge}
\title{Check whether README.md features an rOpenSci software-review badge}
\usage{
has_readme_badge(path = getwd())
}
\arguments{
\item{path}{}
}
\value{
A string, empty if the badge was found.
}
\description{
Check whether README.md features an rOpenSci software-review badge
}
