% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_user_authorized}
\alias{is_user_authorized}
\title{Check whether a user, identified from GitHub API token, is authorized to call
endpoints.}
\usage{
is_user_authorized(secret = NULL)
}
\arguments{
\item{secret}{Environment variable \code{PKGCHECK_TOKEN} sent from bot.}
}
\value{
Logical value indicating whether or not a user is authorized.
}
\description{
This function is used only in the \pkg{plumber} endpoints, to prevent them
being called by unauthorized users.
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{push_to_gh_pages}()},
\code{\link{srr_counts_summary}()},
\code{\link{srr_counts}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
