% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-issue.R
\name{issue_is_stats}
\alias{issue_is_stats}
\title{Determine whether a GitHub issue is a Stats submission}
\usage{
issue_is_stats(orgrepo, issue_num)
}
\arguments{
\item{orgrepo}{GitHub organization and repo as single string separated by
forward slash (\code{org/repo}).}

\item{issue_num}{Number of issue from which to extract submission type.}
}
\value{
\code{TRUE} if the submission type is "Stats", otherwise \code{FALSE}.
}
\description{
Determine whether a GitHub issue is a Stats submission
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{push_to_gh_pages}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts}()},
\code{\link{srr_counts_from_report}()},
\code{\link{srr_counts_summary}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
