% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmatch.R
\name{pkgmatch_repo}
\alias{pkgmatch_repo}
\title{Body of main 'pkgmatch' response}
\usage{
pkgmatch_repo(repourl, repo, issue_id, n_top = 5L, post_to_issue = TRUE)
}
\arguments{
\item{repourl}{The URL for the repo being matched, potentially including
full path to non-default branch.}

\item{repo}{The 'context.repo' parameter defining the repository from which
the command was invoked, passed in 'org/repo' format.}

\item{issue_id}{The id (number) of the issue from which the command was
invoked.}

\item{n_top}{Return this number of top-matches packages.}

\item{post_to_issue}{Integer value > 0 will post results back to issue (via
'gh' cli); otherwise just return character string with result.}
}
\value{
If \code{!post_to_issue}, a markdown-formatted response body, otherwise
URL of the issue comment to which response body has been posted.
}
\description{
Body of main 'pkgmatch' response
}
\seealso{
Other main: 
\code{\link{collate_editor_check}()},
\code{\link{editor_check}()},
\code{\link{serve_api}()}
}
\concept{main}
