% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-to-issue.R
\name{post_to_issue}
\alias{post_to_issue}
\title{Post review checks to GitHub issue}
\usage{
post_to_issue(cmt, repo, issue_id)
}
\arguments{
\item{cmt}{Single character string with comment to post.}

\item{repo}{The repository to post to (obtained directly from bot).}

\item{issue_id}{The number of the issue to post to.}
}
\value{
URL of the comment within the nominated issue
}
\description{
Post review checks to GitHub issue
}
\seealso{
Other github: 
\code{\link{dl_gh_repo}()}
}
\concept{github}
