% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-pages.R
\name{push_to_gh_pages}
\alias{push_to_gh_pages}
\title{Push static \code{html} files to \code{gh-pages} branch of this repo to serve via
GitHub pages.}
\usage{
push_to_gh_pages(check)
}
\arguments{
\item{check}{Return result of \link{editor_check} function.}
}
\value{
Vector of two paths containing URLs of the \code{srr} and \code{network} files.
}
\description{
Push static \code{html} files to \code{gh-pages} branch of this repo to serve via
GitHub pages.
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts}()},
\code{\link{srr_counts_from_report}()},
\code{\link{srr_counts_summary}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
