% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer-review-badge.R
\name{readme_badge}
\alias{readme_badge}
\title{Check whether README.md features an rOpenSci software-review badge}
\usage{
readme_badge(repourl, repo, issue_id, post_to_issue = TRUE)
}
\arguments{
\item{repourl}{The URL for the repo being checked, potentially including full
path to non-default branch.}

\item{repo}{The 'context.repo' parameter defining the repository from which
the command was invoked, passed in 'org/repo' format.}

\item{issue_id}{The id (number) of the issue from which the command was
invoked.}

\item{post_to_issue}{Integer value > 0 will post results back to issue (via
'gh' cli); otherwise just return character string with result.}
}
\value{
A string, empty if the badge was found.
}
\description{
Check whether README.md features an rOpenSci software-review badge
}
