% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated-pkgs.R
\name{rorevapi_updated_pkgs}
\alias{rorevapi_updated_pkgs}
\title{List any direct dependencies which are not latest CRAN versions}
\usage{
rorevapi_updated_pkgs(path)
}
\arguments{
\item{path}{Path to local package}
}
\value{
Character vector of any packages which can be updated.
}
\description{
List any direct dependencies which are not latest CRAN versions
}
\seealso{
Other utils: 
\code{\link{check_cache}()},
\code{\link{pkgrep_install_deps}()},
\code{\link{stdout_stderr_cache}()},
\code{\link{symbol_crs}()},
\code{\link{symbol_tck}()}
}
\concept{utils}
