% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{srr_counts}
\alias{srr_counts}
\title{Count number of 'srr' statistical standards complied with, and confirm
whether than represents > 50\% of all applicable standards.}
\usage{
srr_counts(repourl, repo, issue_id, post_to_issue = TRUE)
}
\arguments{
\item{repourl}{The URL for the repo being checked, potentially including full
path to non-default branch.}

\item{repo}{The 'context.repo' parameter defining the repository from which
the command was invoked, passed in 'org/repo' format.}

\item{issue_id}{The id (number) of the issue from which the command was
invoked.}

\item{post_to_issue}{Integer value > 0 will post results back to issue (via
'gh' cli); otherwise just return character string with result.}
}
\value{
Vector of three numbers:
\enumerate{
\item Number of standards complied with
\item Total number of applicable standards
\item Number complied with as proportion of total
}
}
\description{
Count number of 'srr' statistical standards complied with, and confirm
whether than represents > 50\% of all applicable standards.
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{push_to_gh_pages}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts_summary}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
