% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{srr_counts_from_report}
\alias{srr_counts_from_report}
\title{Extract final counts of 'srr' standards from the report}
\usage{
srr_counts_from_report(srr_rep)
}
\arguments{
\item{srr_rep}{An 'srr' report generated by the \code{srr::srr_report()} function.}
}
\value{
Character vector with markdown-formatted summary summary of numbers
of standards complied with.
}
\description{
Extract final counts of 'srr' standards from the report
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{push_to_gh_pages}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts}()},
\code{\link{srr_counts_summary}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
