% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{stats_badge}
\alias{stats_badge}
\title{Get stats badge grade and standards version for a submission}
\usage{
stats_badge(repo = "ropensci/software-review", issue_num = 258)
}
\arguments{
\item{repo}{The submission repo}

\item{issue_num}{GitHub issue number of submission}
}
\value{
A single character containing the label used directly for the issue
badge
}
\description{
Get stats badge grade and standards version for a submission
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{push_to_gh_pages}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts_summary}()},
\code{\link{srr_counts}()}
}
\concept{ropensci}
