% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{stats_version}
\alias{stats_version}
\title{Get current version of statistical standards}
\usage{
stats_version()
}
\value{
A single character containing version number
}
\description{
Get current version of statistical standards
}
