% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{stdout_stderr_cache}
\alias{stdout_stderr_cache}
\title{Set up stdout & stderr cache files for \code{r_bg} process}
\usage{
stdout_stderr_cache(repourl)
}
\arguments{
\item{repourl}{The URL of the repo being checked}
}
\value{
Vector of two strings holding respective local paths to \code{stdout} and
\code{stderr} files for \code{r_bg} process controlling the main \link{editor_check}
function.
}
\description{
Set up stdout & stderr cache files for \code{r_bg} process
}
\note{
These files are needed for the \pkg{callr} \code{r_bg} process which
controls the main \link{editor_check} call. See
\url{https://github.com/r-lib/callr/issues/204}. The \code{stdout} and \code{stderr}
pipes from the process are stored in the cache directory so they can be
inspected via their own distinct endpoint calls.
}
