% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symbol_crs}
\alias{symbol_crs}
\title{Cross symbol, exported for direct use in plumber API}
\usage{
symbol_crs()
}
\description{
Cross symbol, exported for direct use in plumber API
}
\seealso{
Other utils: 
\code{\link{check_cache}()},
\code{\link{pkgrep_install_deps}()},
\code{\link{rorevapi_updated_pkgs}()},
\code{\link{stdout_stderr_cache}()},
\code{\link{symbol_tck}()}
}
\concept{utils}
