% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_is_r_pkg}
\alias{url_is_r_pkg}
\title{Check whether a given GitHub URL contains an R package.}
\usage{
url_is_r_pkg(repourl)
}
\arguments{
\item{repourl}{Potentially with "/tree/branch_name/sub-directory" appended}
}
\value{
\code{TRUE} if 'repourl' is an R package, otherwise \code{FALSE}
}
\description{
Check whether a given GitHub URL contains an R package.
}
\seealso{
Other github: 
\code{\link{dl_gh_repo}()},
\code{\link{get_branch_from_url}()},
\code{\link{get_subdir_from_url}()},
\code{\link{post_to_issue}()}
}
\concept{github}
