% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-issue.R
\name{check_issue_template}
\alias{check_issue_template}
\title{Check template variables in GitHub issue}
\usage{
check_issue_template(org, repo, issue_num)
}
\arguments{
\item{org}{GitHub organization}

\item{repo}{GitHub repository}

\item{issue_num}{Number of issue from which to extract opening comment}
}
\value{
Comment as character string
}
\description{
Check template variables in GitHub issue
}
