% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor-check.R
\name{file_pkgcheck_issue}
\alias{file_pkgcheck_issue}
\title{File an issue in the GitHub \code{pkgcheck} repo on any packages which fail checks}
\usage{
file_pkgcheck_issue(
  repourl = NULL,
  repo = "ropensci/software-review",
  issue_id = NULL
)
}
\description{
File an issue in the GitHub \code{pkgcheck} repo on any packages which fail checks
}
\note{
Exported only for access by plumber; not intended for general external
usage.
}
