% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{pkgrep_install_deps}
\alias{pkgrep_install_deps}
\title{Install all system and package dependencies of an R package}
\usage{
pkgrep_install_deps(path, os, os_release)
}
\arguments{
\item{path}{Path to local package}

\item{os}{Name of operating system, passed to \pkg{remotes} package to
install system dependencies.}

\item{os_release}{Name of operating system release, passed to \pkg{remotes}
package to install system dependencies.}
}
\value{
Hopefully a character vector of length zero, otherwise a list of any
R packages unable to be installed.
}
\description{
Install all system and package dependencies of an R package
}
\seealso{
Other utils: 
\code{\link{check_cache}()},
\code{\link{stdout_stderr_cache}()},
\code{\link{symbol_crs}()},
\code{\link{symbol_tck}()}
}
\concept{utils}
