% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer-review-badge.R
\name{readme_has_peer_review_badge}
\alias{readme_has_peer_review_badge}
\title{Check whether 'README.md' has a "peer reviewed" badge}
\usage{
readme_has_peer_review_badge(path = getwd(), issue_id = NULL)
}
\arguments{
\item{path}{Local path to package directory.}

\item{issue_id}{The id (number) of the issue from which the command was
invoked.}
}
\value{
A string, empty if the badge was found.
}
\description{
Check whether 'README.md' has a "peer reviewed" badge
}
