% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roreviewapi-package.R
\docType{package}
\name{roreviewapi-package}
\alias{roreviewapi}
\alias{roreviewapi-package}
\title{roreviewapi: Plumber API to report package structure and function}
\description{
Plumber API to report package structure and function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenscilabs.github.io/roreviewapi}
  \item \url{https://github.com/ropenscilabs/roreviewapi}
  \item Report bugs at \url{https://github.com/ropenscilabs/roreviewapi/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
