% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{srr_counts_summary}
\alias{srr_counts_summary}
\title{Summarise counts of 'srr' standards from full 'srr' report}
\usage{
srr_counts_summary(srr_rep)
}
\arguments{
\item{srr_rep}{An 'srr' report generated by the \code{srr::srr_report()} function.}
}
\value{
Character vector with markdown-formatted summary summary of numbers
of standards complied with.
}
\description{
Summarise counts of 'srr' standards from full 'srr' report
}
