% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srr.R
\name{srr_counts_summary}
\alias{srr_counts_summary}
\title{Summarise counts of 'srr' standards from full 'srr' report}
\usage{
srr_counts_summary(srr_rep, has_errors = FALSE)
}
\arguments{
\item{srr_rep}{An 'srr' report generated by the \code{srr::srr_report()} function.}

\item{has_errors}{This is \code{TRUE} when the 'srr' report has errors, in which
case the final message issued here excludes statements about package being
able to be submitted.}
}
\value{
Character vector with markdown-formatted summary summary of numbers
of standards complied with.
}
\description{
Summarise counts of 'srr' standards from full 'srr' report
}
\seealso{
Other ropensci: 
\code{\link{check_issue_template}()},
\code{\link{is_user_authorized}()},
\code{\link{push_to_gh_pages}()},
\code{\link{readme_has_peer_review_badge}()},
\code{\link{srr_counts}()},
\code{\link{stats_badge}()}
}
\concept{ropensci}
